backend/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
Django==5.2.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
#psycopg2-binary==2.9.11 #splite3으로 임시 작업해서 주석 처리해놨습니다
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.1.0
requests==2.32.4
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
